AgregarDispositivo actualizado, todavia no borre el viejo addDispo...

+++ src/main/java/proyectoDDSs/DispositivoEstandar.java
@@ -4,7 +4,7 @@ public class DispositivoEstandar extends Dispositivo {
 	
 	int usoDiarioEnHoras;
 
-	public DispositivoEstandar(String unNombre, long electricidadQConsume, int unaCantidadDeHoras) {
+	public DispositivoEstandar(String unNombre, double electricidadQConsume, int unaCantidadDeHoras) {
 		super(unNombre ,electricidadQConsume);
 		usoDiarioEnHoras = unaCantidadDeHoras;
 	}
